Refactor record types to live in separate source file
